chore: rename session crate
