--- src/day_01/mod.rs
@@ -1,2 +1,0 @@
-pub mod enums;
-pub mod structs;
