+++ src/session_three/functions/attendance_struct.rs
