+++ src/session_three/functions/course_registry.rs
