+++ src/session_three/functions/student_registry.rs
