+++ src/session_three/student_registry.rs
