+++ src/session_three/types/attendance.rs
