+++ src/session_three/types/attendance_status.rs
