+++ src/session_three/types/course_registry.rs
