+++ src/session_three/types/student_registry.rs
