+++ src/session_two/vectors.rs
